Add files via upload
These are the files I used in the code I submitted. The only file not here is a shapefile for school locations at the beginning of the code, as currently I don't know if that's usable. The school data is from 2013-2014, however I could not find anything later from the Chicago Data Portal except shapefiles.