Add files via upload
This contains all of the CPS information with all of the variables used. Each dataset is for the specified year. ES_SB stands for Elementary School with School Boundaries. So the school boundary number has been added to the data set so you can identify which school boundary that particular school is in. HS_SB stands for High School with School Boundary. I will also upload a variable description shortly.