Add files via upload
These files contain the crime data for that specific school year. So if it says 2010, the dates of the crimes are from Aug. 1st 2010 to July 31st 2011. If it says 2011, the dates of the crimes are from Aug. 1st 2011 to July 31st 2012 and so on. Again ES_SB stands for Elementary School with School Boundaries. Therefore each crime in the dataset has a specific school boundary so you know where the crime was located. HOWEVER, there are some crimes that did not happen within any of the school boundaries, usually they took place at O'Hare or they are right outside the school boundaries. So some data points have an NA in the school boundary. Again HS_SB stands for High School with School Boundary.